Update install_elk_0.5_protege.md
The new Elk 0.5 plug-in was named something different than the Elk 0.43 plug-in that I was replacing, so when putting into the plugins directory it did not ask to replace. I had to manually remove the old plugin before the new plugin showed up in the Protege Reasoner menu.  My old plugin was called "org.semanticweb.elk.jar" Also surprisingly, when I tried method 1, which I've done for other programs, this location appears to be different contents than what I saw when I used approach 2 through terminal.  I have screenshots, if interested to see this. It works now that I've removed the old plugin.