fixed typo, made some phrasing parallel (#422)
i came here to fix the typo "an GitHub" but while i was at it, i changed a few words in that section of the document to make the bullet points more parallel. note that there are many other places i did not update, but that would benefit from a similar unification of style -- for example, in the section before the one i edited, there are points "Updating Imports" and "Import ORCIDIO". the first of those could be rephrased "Update imports" or the second could be "Importing ORCIDIO".