Improvements to documentation about how to review PRs (#423)
* Improvements to documentation about how to review PRs

First, I continue to be amazed at what a great resource the OBOOK is. In this PR, I have suggested a few grammatical fixes and wording improvements and some minor reorganization of some content.

A few things:
1. I'm not sure if I made the correct relative URL here:
[here](/howto/github-create-pull-request/)
I tried and failed to find an example of a similar relative reference in an obook module. Please correct as needed.

2. I understand that the point of the "Overarching" section is to summarize the main points of the more detailed "How to" section, but in fact the Overaching section goes into so much detail, it's largely redundant with the How to section. I can't quickly fix that, but I did move the "manual testing" section from the overview down to the more detailed how to section.

3. "in particular line ending changes" -- not sure what these are; please briefly define

* Update docs/howto/review-pull-request.md

---------

Co-authored-by: Nico Matentzoglu <nicolas.matentzoglu@gmail.com>